Refactoring of adaptive time stepping
This PR refactors the adaptive time stepping in preparation of the introduction of a TimeDiscretizationInterface in #25.
For one, this makes error estimation and and error control more independent from one another, and also makes hidden internal data external as part of preparations of introducing a TimeIntegrationState that contains said internal data. This has the additional advantage of making both easier to unit test.
Furthermore, the error estimator class, which previously computed the error estimate and norm the create a measure for the error is reafactored. The norm calculation of the error estimate is now part of the DiscretizedODE, and the (now renamed from ErrorEstimator) ErrorMeasurer is responsible to compute a measure of thereof.